+++ NEWS.md
@@ -3,7 +3,10 @@ title: "Release notes for `IEATools`"
 output: html_document
 ---
 
-
+* `form_eta_fu_phi_u_vecs()` now changes the name of the single column 
+  of the exergy-to-energy ratio vector (`phi_u`) from "phi.u" to simply "phi", 
+  to become compatible with the `phi_pf` vector, whose single column is named "phi".
+  This change fixes a downstream bug in the `SEAPSUTWorkflow` pacakge.
 * Added functions to assist converting energy to exergy
   when data are in PSUT matrix format.
   New functions include: